\name{PEN-package}
\alias{PEN-package}
\docType{package}
\title{Penalized linear regression and logistic regression.}
\description{
 The package is for penalized linear and logistic regression with SCAD, MCP, Log and SICA penalties.
 The structure of the codes is based on ncvreg package, which is developed for SCAD and MCP penalties.
 }
\details{ This package aims to facilitate genome-wide association study. Based on our work on penalty functions study,
we believe that logistic regression or linear regression with Log or SICA penalties has potential to select a set of informative SNPs for the
binary response such as disease status or for the continuous response.}
\author{
Ting-Huei Chen
Maintainer: Ting-Huei Chen <tingstat22@gmail.com>
}
\references{
Breheny, P. and Huang, J. (2011) Coordinate descent
  algorithms for nonconvex penalized regression, with applications to
  biological feature selection. Ann. Appl. Statist., 5: 232-253.
}
