\name{XData}
\alias{XData}
\docType{data}
\title{XData is a covariate matrix.
  
}
\description{
  This data object is used to demonstrate function PEN. XData is a 300 by 1101 matrix simulated as aga covariate plus 1100 SNPs on chromosomes 1 to 22 with 50 SNPs on each chromosome by GWAsimulator (Li and Li (2008)).
}
\keyword{datasets}
