\name{XIndex}
\alias{XIndex}
\docType{data}
\title{ Indicator vector to specify which covariates to be penalized.
  
}
\description{
  This data object is used to demonstrate function PEN. The index vector XIndex with length of number of covariates consists 0 and 1 to index covariates to be unpenalized and penalized respectively.

}
\keyword{datasets}
